Add unit tests outlining the procedure for pickling objects whose C++ side are implemented via composition.
